split leader functions and follower/candidate functions to different files
