+++ pkg/raft/node_test.go
@@ -258,19 +258,19 @@ func TestReplicateLogsTo(t *testing.T) {
 	}
 }
 
-func TestWonMajorityVotes(t *testing.T) {
+func TestWonElection(t *testing.T) {
 	n := &node{}
 	n.clusterSize = 3
 	n.votes = make(map[int]bool)
 
 	n.votes[0] = true
-	if n.wonMajorityVotes() {
-		t.Error("wonMajorityVotes should return false on 1 vote out of 3")
+	if n.wonElection() {
+		t.Error("wonElection should return false on 1 vote out of 3")
 	}
 
 	n.votes[2] = true
-	if !n.wonMajorityVotes() {
-		t.Error("wonMajorityVotes should return true on 2 votes out of 3")
+	if !n.wonElection() {
+		t.Error("wonElection should return true on 2 votes out of 3")
 	}
 
 }
